create SearchContext module since ShadowRoot is not an Element
